# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quire Server API",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "description": "Private reading-data sync and EPUB storage. Account and watched-folder administration are owner CLI operations."
  },
  "paths": {
...
    "/v1/me/password": {
      "put": {
        "summary": "Put me password",
        "tags": [
          "Accounts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed; GET returns the requested representation."
          },
          "204": {
            "description": "Mutation completed without a response body."
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Administrator permission required."
          },
          "409": {
            "description": "Conflicting state."
          }
        }
      }
    },
    "/v1/setup": {
      "get": {
        "summary": "Get setup",
        "tags": [
          "Accounts"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Request completed; GET returns the requested representation."
          },
          "204": {
            "description": "Mutation completed without a response body."
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Administrator permission required."
          },
          "409": {
            "description": "Conflicting state."
          }
        }
      },
      "post": {
        "summary": "Create initial administrator",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "username",
                  "password",
                  "code"
                ],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "code": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created; sign in separately."
          },
          "400": {
            "description": "Invalid fields or expired/redeemed invitation."
          },
          "401": {
            "description": "Invalid setup code."
          },
          "409": {
            "description": "Setup already completed or username unavailable."
          },
          "429": {
            "description": "Retry later."
          }
        }
      }
    },
    "/v1/me": {
      "get": {
        "summary": "Get me",
        "tags": [
          "Accounts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed; GET returns the requested representation."
          },
          "204": {
            "description": "Mutation completed without a response body."
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Administrator permission required."
          },
          "409": {
            "description": "Conflicting state."
          }
        }
      }
    },
    "/v1/admin/users": {
      "get": {
        "summary": "Get admin users",
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed; GET returns the requested representation."
          },
          "204": {
            "description": "Mutation completed without a response body."
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Administrator permission required."
          },
          "409": {
            "description": "Conflicting state."
          }
        },
        "description": "Requires an active administrator account. Does not grant access to another user's annotations or reading positions."
      }
    },
    "/v1/admin/users/{id}": {
      "put": {
        "summary": "Put admin users {id}",
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed; GET returns the requested representation."
          },
          "204": {
            "description": "Mutation completed without a response body."
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Administrator permission required."
          },
          "409": {
            "description": "Conflicting state."
          }
        },
        "description": "Requires an active administrator account. Does not grant access to another user's annotations or reading positions.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/admin/users/{id}/sessions": {
      "delete": {
        "summary": "Delete admin users {id} sessions",
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed; GET returns the requested representation."
          },
          "204": {
            "description": "Mutation completed without a response body."
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Administrator permission required."
          },
          "409": {
            "description": "Conflicting state."
          }
        },
        "description": "Requires an active administrator account. Does not grant access to another user's annotations or reading positions.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/admin/invites": {
      "post": {
        "summary": "Post admin invites",
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed; GET returns the requested representation."
          },
          "204": {
            "description": "Mutation completed without a response body."
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Administrator permission required."
          },
          "409": {
            "description": "Conflicting state."
          },
          "201": {
            "description": "Resource created."
          }
        },
        "description": "Requires an active administrator account. Does not grant access to another user's annotations or reading positions."
      },
      "get": {
        "summary": "Get admin invites",
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed; GET returns the requested representation."
          },
          "204": {
            "description": "Mutation completed without a response body."
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Administrator permission required."
          },
          "409": {
            "description": "Conflicting state."
          }
        },
        "description": "Requires an active administrator account. Does not grant access to another user's annotations or reading positions."
      }
    },
    "/v1/admin/invites/{id}": {
      "delete": {
        "summary": "Delete admin invites {id}",
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed; GET returns the requested representation."
          },
          "204": {
            "description": "Mutation completed without a response body."
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Administrator permission required."
          },
          "409": {
            "description": "Conflicting state."
          }
        },
        "description": "Requires an active administrator account. Does not grant access to another user's annotations or reading positions.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}